.aux/数据库-哈利波特/embedding/requirements.txt
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pyyaml>=6.0
regex>=2024.0
